Move email sending generic method to common
